+++ pkg/streamingpromql/query.go
@@ -493,7 +493,10 @@ func (q *Query) Exec(ctx context.Context) *promql.Result {
 		return &promql.Result{Err: compat.NewNotSupportedError(fmt.Sprintf("unsupported result type %s", parser.DocumentedType(q.statement.Expr.Type())))}
 	}
 
-	q.result.Warnings = *q.annotations
+	// To make comparing to Prometheus' engine easier, only return the annotations if there are some, otherwise, return nil.
+	if len(*q.annotations) > 0 {
+		q.result.Warnings = *q.annotations
+	}
 
 	return q.result
 }
@@ -569,6 +572,10 @@ func (q *Query) populateMatrixFromInstantVectorOperator(ctx context.Context, o t
 			Floats:     d.Floats,
 			Histograms: d.Histograms,
 		})
+	}
+
+	if len(m) == 0 {
+		return nil, nil
 	}
 
 	slices.SortFunc(m, func(a, b promql.Series) int {